+++ src/renderer/components/memes/nc.tsx
@@ -1,6 +1,7 @@
-import { stealInput } from '../../core/input'
+//import { stealInput } from '../../core/input'
 import { render } from 'inferno'
-import api from '../../core/instance-api'
+import { Events } from '../../../common/ipc'
+// import api from '../../core/instance-api'
 
 let state = {
   visible: false,
@@ -36,10 +37,11 @@ const hide = () => {
   }
 }
 
-api.onAction('nc', () => {
+window.api.on(Events.ncAction, () => {
   show()
-  const restoreInput = stealInput(() => {
-    restoreInput()
+  // TODO(smolck): Need to `await` any of this?
+  window.api.stealInput(() => {
+    window.api.restoreInput()
     hide()
   })
 })
