+++ src/renderer/components/nvim/autocomplete.tsx
@@ -1,12 +1,12 @@
 import { RowNormal } from '../row-container'
 import { resetMarkdownHTMLStyle } from '../../ui/styles'
 import * as windows from '../../windows/window-manager'
-import * as dispatch from '../../messaging/dispatch'
-import * as workspace from '../../core/workspace'
+import * as dispatch from '../../dispatch'
+import * as workspace from '../../workspace'
 import { PopupMenu } from '../../render/events'
 import { paddingVH, cvar } from '../../ui/css'
 import Overlay from '../overlay'
-import { cursor } from '../../core/cursor'
+import { cursor } from '../../cursor'
 import { parse as stringToMarkdown } from 'marked'
 import { render } from 'inferno'
 import Icon from '../icon'
